Use view files for generator commands
