+++ system/Commands/Generators/CreateSessionMigration.php
@@ -104,7 +104,8 @@ class CreateSessionMigration extends GeneratorCommand
 			'matchIP'   => config('App')->sessionMatchIP ?? false,
 		];
 
-		$template = view('\\CodeIgniter\\Commands\\Generators\\Views\\migration.tpl.php', $data, ['debug' => false]);
-		return str_replace('@php', '?php', $template);
+		$template = view('CodeIgniter\\Commands\\Generators\\Views\\session_migration.tpl.php', $data, ['debug' => false]);
+
+		return str_replace('<@php', '<?php', $template);
 	}
 }
