+++ system/Commands/Generators/Views/seed.tpl.php
@@ -1,0 +1,13 @@
+<@php
+
+namespace {namespace};
+
+use CodeIgniter\Database\Seeder;
+
+class {class} extends Seeder
+{
+	public function run()
+	{
+		//
+	}
+}
